✨ Adiciona tabelas de documentos pendentes e devidos ao componente Espaço Gerencial
