+++ src/app/comercial/espaco/pending-docs-table/columns.tsx
@@ -15,6 +15,7 @@ import { SheetViewComercial } from '@/components/Features/Comercial/SheetViewCom
 import { RiSidebarUnfoldLine } from 'react-icons/ri';
 import { CoordinatorParticipationChart } from '@/components/Charts/CommissionParticipationChart';
 import { ComercialContext } from '@/context/ComercialContext';
+import { IDocTable } from '@/interfaces/IPendingDocResponse';
 
 const CellComponent = (row: { original: any }) => {
     const { setSheetOpen, sheetOpen, setSheetOpenId } = React.useContext(ComercialContext);
@@ -35,24 +36,7 @@ const CellComponent = (row: { original: any }) => {
     );
 };
 
-export const columns: ColumnDef<ITabelaGerencial>[] = [
-    {
-        accessorKey: 'criado_em',
-        accessorFn: (row) => dateFormater(row.criado_em.split('T')[0]),
-        header: ({ column }) => {
-            return (
-                <Button
-                    variant={'ghost'}
-                    className="flex max-w-20 items-center gap-2"
-                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-                >
-                    Criado em
-                    <ArrowUpDown size={16} />
-                </Button>
-            );
-        },
-        cell: ({ row }) => <div className="max-w-20 lowercase">{row.getValue('criado_em')}</div>,
-    },
+export const columnsPendingDoc: ColumnDef<IDocTable>[] = [
     {
         accessorKey: 'usuario',
         header: ({ column }) => {
@@ -74,147 +58,243 @@ export const columns: ColumnDef<ITabelaGerencial>[] = [
         ),
     },
     {
-        accessorKey: 'credor',
-        header: () => {
-            return <span className="flex min-w-2 max-w-36 items-center gap-4">Nome do Credor</span>;
-        },
-        cell: ({ row }) => (
-            <div
-                className="flex max-w-90 items-center overflow-auto text-nowrap"
-                style={{
-                    scrollbarWidth: 'thin',
-                    scrollbarColor: 'rgba(0, 0, 0, 0.2) rgba(0, 0, 0, 0.1)',
-                }}
-            >
-                <span className="">{row.getValue('credor')}</span>
-                <CellComponent original={row.original} />
-            </div>
-        ),
-    },
-    // {
-    //     accessorKey: 'sheet',
-    //     header: ({ column }) => {
-    //         return null;
-    //     },
-    //     cell: ({ row }) => <CellComponent original={row.original} />,
-    // },
-    // {
-    //     accessorKey: 'observacoes',
-    //     header: ({ column }) => {
-    //         return (
-    //             <Button
-    //                 variant="ghost"
-    //                 className="flex justify-start gap-2"
-    //                 onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-    //             >
-    //                 <p className="flex justify-start">Observações</p>
-    //                 <ArrowUpDown size={16} />
-    //             </Button>
-    //         );
-    //     },
-    //     cell: ({ row }) => {
-    //         return <p className="max-w-36 truncate">{row.getValue('observacoes')}</p>;
-    //     },
-    // },
-    // {
-    //     header: 'Comissão',
-    //     cell: ({ row }) => {
-    //         const resumo = row.original;
-    //         return (
-    //             <div className="flex items-center">
-    //                 <span>{numberFormat(resumo.comissao)}</span>
-    //             </div>
-    //         );
-    //     },
-    // },
-    // {
-    //     accessorKey: 'proposta_escolhida',
-    //     header: 'Proposta',
-    //     cell: ({ row }) => (
-    //         <div className="flex items-center">
-    //             <span>{numberFormat(row.getValue('proposta_escolhida'))}</span>
-    //         </div>
-    //     ),
-    // },
-    {
-        accessorKey: 'valor_liquido_disponivel',
-        header: 'Valor Líquido',
-        cell: ({ row }) => (
-            <div className="flex items-center">
-                <span>{numberFormat(row.getValue('valor_liquido_disponivel'))}</span>
-            </div>
-        ),
-    },
-    {
-        accessorKey: 'custo_do_precatorio',
+        accessorKey: 'nome',
         header: ({ column }) => {
             return (
                 <Button
-                    variant="ghost"
-                    className="flex max-w-18 items-center gap-2"
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
                     onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
                 >
-                    Custo do Precatório
+                    Nome
                     <ArrowUpDown size={16} />
                 </Button>
             );
         },
         cell: ({ row }) => (
-            <div className="flex items-center">
-                <span>{percentageFormater(row.getValue('custo_do_precatorio'))}</span>
-            </div>
-        ),
-    },
-
-    // {
-    //     accessorKey: 'proposta_minima',
-    //     header: ({ column }) => {
-    //         return (
-    //             <Button
-    //                 variant="ghost"
-    //                 className="flex min-w-2 max-w-36 items-center gap-4"
-    //                 onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
-    //             >
-    //                 Proposta Mínima
-    //                 <ArrowUpDown size={16} />
-    //             </Button>
-    //         );
-    //     },
-    //     cell: ({ row }) => (
-    //         <div className="flex items-center">
-    //             <span>{numberFormat(row.getValue('proposta_minima'))}</span>
-    //         </div>
-    //     ),
-    // },
-    {
-        accessorKey: 'status',
-        header: ({ column }) => {
-            return <p className="max-w-30 items-center gap-2 truncate">Status</p>;
-        },
-        cell: ({ row }) => (
-            <div className="flex items-center">
-                <p className="max-w-30 truncate">{row.getValue('status')}</p>
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('nome')}</p>
             </div>
         ),
     },
     {
         accessorKey: 'status_diligencia',
         header: ({ column }) => {
-            return <p className="max-w-30 items-center gap-2 truncate">Status Diligência</p>;
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Status Diligência
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
         },
         cell: ({ row }) => (
-            <div className="flex items-center">
-                <p className="max-w-30 truncate">{row.getValue('status_diligencia')}</p>
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status_diligencia')}</p>
             </div>
         ),
     },
-    // {
-    //   id: "actions",
-    //   enableHiding: false,
-    //   cell: ({ row }) => {
-    //     return (
-    //       <CellComponent row={row} />
-    //     )
-    //   },
-    // },
+    {
+        accessorKey: 'status',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Status
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'cpf',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    CPF
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('cpf')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'rg',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    RG
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('rg')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'status_rg',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                   Status RG
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status_rg')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'certidao_nasc_cas',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Certidão de Nascimento/Casamento
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('certidao_nasc_cas')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'status_certidao_nasc_cas',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Status Certidão de Nascimento/Casamento
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status_certidao_nasc_cas')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'comprovante_residencia',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Comprovante de Residência
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('comprovante_residencia')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'status_comprovante_residencia',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Status Comprovante de Residência
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status_comprovante_residencia')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'oficio_requisitorio',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Ofício Requisitório
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('oficio_requisitorio')}</p>
+            </div>
+        ),
+    },
+    {
+        accessorKey: 'status_oficio_requisitorio',
+        header: ({ column }) => {
+            return (
+                <Button
+                    variant={'ghost'}
+                    className="flex items-center gap-2"
+                    onClick={() => column.toggleSorting(column.getIsSorted() === 'asc')}
+                >
+                    Status Ofício Requisitório
+                    <ArrowUpDown size={16} />
+                </Button>
+            );
+        },
+        cell: ({ row }) => (
+            <div className="w-33 gap-2">
+                <p className="truncate">{row.getValue('status_oficio_requisitorio')}</p>
+            </div>
+        ),
+    },
 ];
